+++ src/tools/flightplan/show-aircraft-list.ts
@@ -32,7 +32,6 @@ export async function ShowAircraftList() {
 	if (!editor) {
 		return;
 	}
-	const config = vscode.workspace.getConfiguration('fs-ai-tools.showAircraftList', undefined);
 
 	const filePath = editor.document.uri.path;
 	const dirPath = Path.dirname(filePath).replace(/^\/+/, '');
@@ -64,8 +63,11 @@ export async function ShowAircraftList() {
 	const { aircraftList, totalCount, nonMatches } = matchTitleToType(aircraftListRaw);
 
 	// 4. Show formatted message with "copy" button
+	const showGoogleSheetsButton = vscode.workspace
+		.getConfiguration('fs-ai-tools.showAircraftList', undefined)
+		.get('showGoogleSheetsButton');
 	const formattedList = getFormattedAircraftList(aircraftList, totalCount, nonMatches);
-	if (config.showGoogleSheetsButton) {
+	if (showGoogleSheetsButton) {
 		vscode.window
 			.showInformationMessage(formattedList, { modal: true }, 'Copy for Google Sheets')
 			.then((buttonText) => {
@@ -276,7 +278,15 @@ function generateGoogleSheetsOutput(aircraftList: aircraftList) {
  * Iterates through the aircraftList items to provide a single formatted, readable string with "{type}: {count}× ({number of variations})"
  */
 function getFormattedAircraftList(aircraftList: aircraftList, totalCount: number, nonMatches: string[]): string {
-	const title = `${'—'.repeat(10)}   ${totalCount} aircraft   ${'—'.repeat(10)}`;
+	const dialogStyle = vscode.workspace.getConfiguration('window', undefined).get('dialogStyle');
+
+	let title: string;
+	if (dialogStyle === 'custom') {
+		title = `${'—'.repeat(10)}   ${totalCount} aircraft   ${'—'.repeat(10)}`;
+	} else {
+		title = `${totalCount} aircraft:`;
+	}
+
 	const output: string[] = [title, ''];
 	aircraftList.forEach((data, key) => {
 		let text = `• ${data.name || key}: ${data.count}×`;
@@ -289,13 +299,13 @@ function getFormattedAircraftList(aircraftList: aircraftList, totalCount: number
 	if (nonMatches.length) {
 		output.push(
 			'',
-			'—'.repeat(25),
-			`${'aircraft title'.plural(nonMatches.length)} couldn't be matched to any aircraft type:`
+			dialogStyle === 'custom' ? '—'.repeat(25) : '',
+			`❌ ${'aircraft title'.plural(nonMatches.length)} couldn't be matched to any aircraft type:`
 		);
 		output.push(...nonMatches.map((title) => `• "${title}"`));
 	}
 
-	return output.join('\n');
+	return output.join(dialogStyle === 'custom' ? '\n' : ' ');
 }
 
 function searchTermIsRegex(term: string): boolean {